requirements.txt
appdirs==1.4.4 ; python_version >= "3.8" and python_version < "3.9"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "3.9"
cffi==1.15.1 ; python_version >= "3.8" and python_version < "3.9"
charset-normalizer==3.1.0 ; python_version >= "3.8" and python_version < "3.9"
click==8.1.3 ; python_version >= "3.8" and python_version < "3.9"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "3.9"
commonmark==0.9.1 ; python_version >= "3.8" and python_version < "3.9"
cryptography==39.0.2 ; python_version >= "3.8" and python_version < "3.9"
cycler==0.11.0 ; python_version >= "3.8" and python_version < "3.9"
docker-pycreds==0.4.0 ; python_version >= "3.8" and python_version < "3.9"
docstring-parser==0.15 ; python_version >= "3.8" and python_version < "3.9"
einops==0.6.0 ; python_version >= "3.8" and python_version < "3.9"
ffmpeg==1.4 ; python_version >= "3.8" and python_version < "3.9"
ffprobe==0.5 ; python_version >= "3.8" and python_version < "3.9"
gitdb==4.0.10 ; python_version >= "3.8" and python_version < "3.9"
gitpython==3.1.31 ; python_version >= "3.8" and python_version < "3.9"
idna==3.4 ; python_version >= "3.8" and python_version < "3.9"
imageio==2.26.1 ; python_version >= "3.8" and python_version < "3.9"
kiwisolver==1.4.4 ; python_version >= "3.8" and python_version < "3.9"
loguru==0.6.0 ; python_version >= "3.8" and python_version < "3.9"
matplotlib==3.3.4 ; python_version >= "3.8" and python_version < "3.9"
mypy-extensions==1.0.0 ; python_version >= "3.8" and python_version < "3.9"
networkx==3.0 ; python_version >= "3.8" and python_version < "3.9"
numpy==1.24.2 ; python_version >= "3.8" and python_version < "3.9"
pandas==1.5.3 ; python_version >= "3.8" and python_version < "3.9"
pathtools==0.1.2 ; python_version >= "3.8" and python_version < "3.9"
pillow==9.4.0 ; python_version >= "3.8" and python_version < "3.9"
protobuf==4.22.1 ; python_version < "3.9" and python_version >= "3.8"
psutil==5.9.4 ; python_version >= "3.8" and python_version < "3.9"
pycparser==2.21 ; python_version >= "3.8" and python_version < "3.9"
pygments==2.14.0 ; python_version >= "3.8" and python_version < "3.9"
pyparsing==3.0.9 ; python_version >= "3.8" and python_version < "3.9"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "3.9"
pytz==2022.7.1 ; python_version >= "3.8" and python_version < "3.9"
pywavelets==1.4.1 ; python_version >= "3.8" and python_version < "3.9"
pyyaml==6.0 ; python_version >= "3.8" and python_version < "3.9"
requests==2.28.2 ; python_version >= "3.8" and python_version < "3.9"
rich==10.16.2 ; python_version >= "3.8" and python_version < "3.9"
scikit-image==0.17.2 ; python_version >= "3.8" and python_version < "3.9"
scikit-video==1.1.11 ; python_version >= "3.8" and python_version < "3.9"
scipy==1.10.1 ; python_version >= "3.8" and python_version < "3.9"
sentry-sdk==1.17.0 ; python_version >= "3.8" and python_version < "3.9"
setproctitle==1.3.2 ; python_version >= "3.8" and python_version < "3.9"
setuptools==67.6.0 ; python_version >= "3.8" and python_version < "3.9"
shellingham==1.5.0.post1 ; python_version >= "3.8" and python_version < "3.9"
six==1.16.0 ; python_version >= "3.8" and python_version < "3.9"
smmap==5.0.0 ; python_version >= "3.8" and python_version < "3.9"
tifffile==2023.3.21 ; python_version >= "3.8" and python_version < "3.9"
tqdm==4.65.0 ; python_version >= "3.8" and python_version < "3.9"
typed-argument-parser==1.8.0 ; python_version >= "3.8" and python_version < "3.9"
typer[all]==0.7.0 ; python_version >= "3.8" and python_version < "3.9"
typing-extensions==4.5.0 ; python_version >= "3.8" and python_version < "3.9"
typing-inspect==0.8.0 ; python_version >= "3.8" and python_version < "3.9"
urllib3==1.26.15 ; python_version >= "3.8" and python_version < "3.9"
wandb==0.14.0 ; python_version >= "3.8" and python_version < "3.9"
win32-setctime==1.1.0 ; python_version >= "3.8" and python_version < "3.9" and sys_platform == "win32"
